refactor: 修改 router 以及 controller 分類
1. 移除 member
2. 新增auth, user
