+++ src/routes/v1/index.js
@@ -1,9 +1,10 @@
 const express = require('express');
-const memberRouter = require('./members');
+const authRouter = require('./auth');
+const userRouter = require('./user');
 
 const router = express.Router();
 
-router.use('/auth', memberRouter);
-router.use('/user', memberRouter);
+router.use('/auth', authRouter);
+router.use('/user', userRouter);
 
 module.exports = router;
